Issue #151 Small optimisation for fft to avoid an unnecessary cast. (saves 4% native C)

+++ src/lib/bm_fft/java/javax/rtcbench/RTCBenchmark.java
@@ -152,7 +152,7 @@ public class RTCBenchmark {
 			l = n;
 			do {
 				l >>= 1;
-			} while (mr+l > nn);
+			} while ((short)(mr+l) > nn);
 			mr = (short)((mr & (l-1)) + l);
 			if (mr <= m)
 				continue;
